refactor: share query channel protocol mapping
