+++ src/common/base/Cargo.toml
@@ -21,6 +21,7 @@ lazy_static.workspace = true
 pin-project.workspace = true
 regex.workspace = true
 serde = { version = "1.0", features = ["derive"] }
+strum.workspace = true
 tokio.workspace = true
 zeroize = { version = "1.6", default-features = false, features = ["alloc"] }
