+++ src/common/meta/Cargo.toml
@@ -41,7 +41,6 @@ common-procedure.workspace = true
 common-query.workspace = true
 common-recordbatch.workspace = true
 common-runtime.workspace = true
-common-session.workspace = true
 common-telemetry.workspace = true
 common-time.workspace = true
 common-wal.workspace = true
