+++ src/session/Cargo.toml
@@ -15,6 +15,7 @@ ahash.workspace = true
 api.workspace = true
 arc-swap = "1.5"
 auth.workspace = true
+common-base.workspace = true
 common-catalog.workspace = true
 common-error.workspace = true
 common-macro.workspace = true
